+++ src/main/java/laddergame/user/UserNames.java
@@ -1,4 +1,4 @@
-package ladder.user;
+package laddergame.user;
 
 import java.util.Arrays;
 import java.util.List;
