+++ src/test/java/laddergame/user/UserNameTest.java
@@ -1,36 +1,38 @@
-package ladder.user;
+package laddergame.user;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.MethodSource;
+import org.junit.jupiter.params.provider.ValueSource;
 
 import java.util.stream.Stream;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
-@DisplayName("참여자들 이름은")
-class UserNamesTest {
+@DisplayName("참여자 이름은")
+class UserNameTest {
 
-    @DisplayName("쉼표로 구분 됨")
+    @DisplayName("이름을 생성자의 파라메터로 가짐")
     @Test
     void create() {
-        String names = "pobi,honux,crong,jk";
-        assertThat(new UserNames(names)).isEqualTo(new UserNames(names));
+        assertThat(new UserName("joy")).isEqualTo(new UserName("joy"));
     }
 
-    @DisplayName("쉼표 이외의 구분자 사용시 예외 발생")
-    @Test
-    void delimiterNotComma() {
-        assertThatThrownBy( () -> new UserNames("pobi/crong")).isInstanceOf(IllegalArgumentException.class);
+    @DisplayName("1자 미만 or 5자 초과시 예외 발생")
+    @ParameterizedTest
+    @ValueSource(strings={"", "abcdef"})
+    void nameLength(String name) {
+        assertThatThrownBy(() -> new UserName(name))
+                .isInstanceOf(IllegalArgumentException.class);
     }
 
     @DisplayName("빈 값이면 예외 발생")
     @ParameterizedTest
     @MethodSource("blankString")
     void isBlank(String blankValue) {
-        assertThatThrownBy(() -> new UserNames(blankValue))
+        assertThatThrownBy(() -> new UserName(blankValue))
                 .isInstanceOf(IllegalArgumentException.class);
     }
